+++ src/formatter/Formatter.ts
@@ -1,15 +1,17 @@
 import type { FormatOptions } from 'src/types';
-import Params from './Params';
-import Tokenizer from './Tokenizer';
-import formatCommaPositions from '../formatter/formatCommaPositions';
-import formatAliasPositions from '../formatter/formatAliasPositions';
-import Parser from '../parser/Parser';
-import ExpressionFormatter from '../formatter/ExpressionFormatter';
-import { indentString } from './config';
-import AliasAs from '../formatter/AliasAs';
-import { Statement } from '../parser/ast';
-import Layout, { WS } from '../formatter/Layout';
-import Indentation from '../formatter/Indentation';
+import { indentString } from 'src/core/config';
+import Params from 'src/core/Params';
+import Tokenizer from 'src/core/Tokenizer';
+
+import Parser from 'src/parser/Parser';
+import { Statement } from 'src/parser/ast';
+
+import formatCommaPositions from './formatCommaPositions';
+import formatAliasPositions from './formatAliasPositions';
+import ExpressionFormatter from './ExpressionFormatter';
+import AliasAs from './AliasAs';
+import Layout, { WS } from './Layout';
+import Indentation from './Indentation';
 
 /** Main formatter class that produces a final output string from list of tokens */
 export default class Formatter {
